create characters at skeleton root instead of mesh level, assorted fixes
This adds support for using multiple meshes with the same skin, as well as skeletons with multiple root bones.  It also fixes characters with transformations to properly do the skinning in global space.

When morph targets are encountered under a Character, they are added as sliders to that Character.  Otherwise, a per-mesh Character is created.

When exporting from Blender, this creates the Character at the node where the Armature is in Blender.

Fixes #27