converter: Sample animations instead of assuming one data value per frame
STEP and LINEAR interpolation has been implemented or skeletal mesh
animations and morph targets. CUBICSPLINE is still todo.

Also, we now respect the input accessor associated with a given channel
instead of using one input accessor for all channels targeting a given
path.

Since we are sampling/re-sampling animations, we pick a target fps and
sample the animations to get enough frames of data. This fps is
controlled via GltfSettings.animation_fps, which defaults to 30.